Some changes made.
- Deleted gpt_utils.py as I could integrate the same thing in utility.py. Placed all GPT functions in utility.py

- Fixed issue with links. Please view comments to understand implementation.

- Removed apple internship function since the Github Repo already has those opportunities. Also, it would cause a webscrape overload.
